Bug: Int to float

+++ pg_salience_feature/inference/constrained_beam_search.py
@@ -321,7 +321,7 @@ class ConstrainedBeamSearch:
         finished = torch.zeros((batch_size*self.beam_size, ), device=device)
         is_pad_dist_set = False
         scores_accumulated = torch.zeros((batch_size * self.beam_size, 1), device=device)
-        inactive = torch.zeros((batch_size * self.beam_size), dtype=torch.int32, device=device)
+        inactive = torch.zeros((batch_size * self.beam_size), device=device)
         for t in range(1, self.max_steps):
             # shape: (batch_size, num_classes)
             target_dists, state = step(best_word_indices.type(torch.long), state, True)
